mov instruction has been implemented

+++ registers.h
@@ -7,27 +7,22 @@ using namespace std;
 class registers
 {
 private:
-    int16_t AX, BX, CX, DX;
-    int16_t CS, DS, SS, ES;
-    int16_t SP, BP, SI, DI;
-    int16_t AL, BL, CL, DL;
-    int16_t AH, BH, CH, DH;
-    int16_t flag;
-    map<string, int16_t *> regi;
+    int32_t AX, BX, CX, DX;
+    int32_t CS, DS, SS, ES;
+    int32_t SP, BP, SI, DI;
+    int32_t AL, BL, CL, DL;
+    int32_t AH, BH, CH, DH;
+    int32_t flag;
+    map<string, int32_t *> regi;
 
 public:
     registers();
 
     bool is_segment_register(string &);
-
     bool is_register(string &);
-
-    int16_t get_data(string &);
-
-    void set_value(string &, int16_t);
-
+    int32_t get_data(string &);
+    void set_value(string &, int32_t);
     void print_register_map();
-
     void print_reg_data_hex_format(string &);
 };
